+++ example/d_art/game.dart
@@ -153,8 +153,8 @@ zigZag(d_pen.Pen pen) {
 }
 
 runRandom(d_pen.Pen pen) {
-  //d_pen.randomProgram(pen);
-  d_pen.randomCommandSequence(pen);
+  d_pen.randomProgram(pen);
+  //d_pen.randomCommandSequence(pen);
   //d_pen.randomExample(pen);
   //d_pen.randomDemo(pen);
 }
@@ -163,8 +163,8 @@ runProgram(d_pen.Pen pen) {
   //onako(pen);
   //rotateSquare(pen);
   //stairSquare(pen);
-  barGraph(pen);
-  //mapMap(pen);
+  //barGraph(pen);
+  mapMap(pen);
   //zigZag(pen);
 }