+++ lib/model/grid/grid.dart
@@ -32,7 +32,7 @@ class Grid {
   num get height => table.area.height;
   Size get size => table.size;
   int get columnCount => table.size.columnCount;
-  int get rowCount => table.size.columnCount;
+  int get rowCount => table.size.rowCount;
   num get cellWidth => table.cellWidth;
   num get cellHeight => table.cellHeight;
 
@@ -57,11 +57,6 @@ class MazeCell extends CellPiece {
 //www.migapro.com/depth-first-search
 class MazeGrid extends Grid {
   MazeGrid(Table table): super(table) {
-    if (columnCount.isOdd || rowCount.isOdd) {
-      var msg = 'Maze grid size must be an even integer: (${size.columnCount}, ${size.rowCount})';
-      print(msg);
-      throw new Exception(msg);
-    }
     var sp = randomCellPiece();
     path(sp.cell.column, sp.cell.row);
   }